declare our local env types in new file as yarn build recreates next-env.d.ts https://nextjs.org/docs/basic-features/typescript
